REFACTOR: extract fetchTasks fun
